added base filter test class
